renamed driver routes to post routes, connected signup page to main handlebars
