+++ views/layouts/main.handlebars
@@ -95,7 +95,7 @@
       </div>
       <div class="w3-container">
         <h5>Need to Create an Account?
-          </br> <a href="/signup"><b>SIGN UP HERE</b></a></h5>
+          </br> <a href="/"><b>SIGN UP HERE</b></a></h5>
       </div>
     </div>
   </div>
